CLN : literalinclude instance editor selection
	new file:   docs/source/traitsui_user_manual/examples/instance_editor_selection.py
	modified:   docs/source/traitsui_user_manual/factories_basic.rst
	deleted:    examples/tutorials/doc_examples/examples/instance_editor_selection.py
